yt-ai-extension/backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
langchain-google-genai
faiss-cpu
youtube-transcript-api
yt-dlp
tiktoken
sentence-transformers
python-dotenv
